split tests up by test set
