+++ test/testsets/Examples.jl
@@ -1,0 +1,4 @@
+# TODO may want to test this but it's very slow on azure
+@otestset "Examples" begin
+    @test_all_no_arg_functions Examples
+end # testset Examples
