+++ test/testsets/TestData.jl
@@ -1,0 +1,3 @@
+@otestset "TestData" begin
+    @test_all_no_arg_functions TestData
+end # testset TestData
